requirements.txt
watchdog==4.0.1
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
